Moving date formatting back to the select factory
